Adding a "sketch" of how all of the pieces will work together.

+++ Indexer/Indexer/Searching/IIndexerSearcher.cs
@@ -7,6 +7,6 @@ namespace Sando.Indexer.Searching
 {
 	public interface IIndexerSearcher
 	{
-		List<Tuple<ProgramElement,Double>> Search(SearchCriteria searchCriteria);
+		List<Tuple<ProgramElement,float>> Search(SearchCriteria searchCriteria);
 	}
 }
